% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.match.R
\name{allele.match}
\alias{allele.match}
\title{Allele match}
\usage{
allele.match(hap)
}
\arguments{
\item{hap}{a hap object}
}
\description{
This function compares alleles across lines in a hap object
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}
}
\keyword{alleles,}
\keyword{hap}

