\name{gbs-package}

\alias{gbs-package}

\alias{gbs}

\docType{package}

\title{
Package for Generalized Birnbaum-Saunders Distributions
}

\description{
A collection of utilities for analyzing censored and uncensored data from generalized Birnbaum-Saunders distributions.
}

\details{
\tabular{ll}{
Package: \tab gbs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2008-05-05\cr
License: \tab GPL\cr
}
}


\references{Diaz-Garcia, J.A., Leiva, V. (2005)  A new family of life
distributions based on elliptically contoured distributions. J. Stat. Plan. Infer.
128:445-457 (Erratum: J. Stat. Plan. Infer. 137:1512-1513).

Leiva, V., Barros, M., Paula, G.A., Sanhueza, A. (2008) Generalized 
Birnbaum-Saunders distributions applied to air pollutant concentration.
Environmetrics 19:235-249.

Sanhueza, A., Leiva, V., Balakrishnan, N. (2008) The generalized Birnbaum-Saunders
distribution and its theory, methodology and application. Comm. Stat. Theory and
Meth. 37:645-670.}

\author{
Barros, Michelli <michelli.karinne@gmail.com> \cr 
Leiva, Victor <victor.leiva@uv.cl, victor.leiva@yahoo.com> \cr
Paula, Gilberto A. <giapaula@ime.usp.br> 
}

\keyword{package}
