% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.summary.R
\name{gbs.summary}
\alias{gbs.summary}
\title{summarize the gbs object.}
\usage{
gbs.summary(hap, geno = F, encoding = c(-1, 0, 1), graph = F)
}
\arguments{
\item{hap}{The gbs object to be processed.}

\item{geno}{A logical value that will convert the marker calls to a numeric geno format.}

\item{encoding}{The numbers used to encode alleles (minor, het, major).}

\item{graph}{A logical value that will output graphs for summary statistics (blank wells, maf, percent present, het, percent het).}
}
\description{
Processes the gbs object, calculates allele statistics, returns allele stats, geno object, and graphs several statistics.
}
\details{

}
\examples{
data(wheat)
hap = hap.read(wheat)
hap = filter.summary(hap,project="wheat",output="hap",graph=T)

}
\author{
Trevor Rife, \email{trife@ksu.edu}

Jesse Poland, \email{jpoland@ksu.edu}
}

