% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GBS_Sampling.R
\name{GBS_Sampling}
\alias{GBS_Sampling}
\title{Sampling Function of Generative Bootstrap Sampler}
\usage{
GBS_Sampling(B1=1000, B2=500, B3=50, B10 = 50, level = 3, S=100)
}
\arguments{
\item{B1}{The number of bootstrap samples in the first level}

\item{B2}{The number of bootstrap samples in the second level}

\item{B3}{The number of bootstrap samples in the third level}

\item{B10}{The number of the first level bootstrap samples to be evaluated at each iteration. This is to reduce the amount of RAM and GPU memory in storing the bootstrap samples.}

\item{level}{Bootstrap level; e.g. a single bootstrap = 1 and a double bootstrap = 2. The maximum is 3.}

\item{S}{The number of subgroups in the subgroup bootstrap}
}
\value{
  \item{time_samp}{computation time for sampling}
  \item{level}{ The level of iterated bootstrap}
  \item{S}{The subgroup size}
  \item{hidden_size}{ The number of nodes for each hidden layer}
  \item{L}{The number of hidden layers}
}
\description{
Generate bootstrap sample from the trained generator of GBS.
}
\details{
}
\examples{
#Samples = GBS_Sampling()
}
\seealso{
GBS
}
\author{
Minsuk Shin and Jun Liu
}
