% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno.impute.R
\name{geno.impute}
\alias{geno.impute}
\title{Impute GBS markers.}
\usage{
geno.impute(geno, method, k = 4, maxiter = 10, n.core = 1, ...)
}
\arguments{
\item{geno}{The geno object to impute. Each row is an individual and each column is a marker. Genotypic data is expected to be encoded as \code{-1,0,1} for minor, het, and major, respectively.}

\item{method}{The method to use for imputation. See details.}

\item{k}{The number of clusters for KNN imputation.}

\item{maxiter}{The maximum number of iterations for RF and SVD imputation.}

\item{n.core}{The number of cores to use for RF imputation. This option is not available on Windows.}

\item{...}{Additional arguments for rrBLUP::A.mat, randomForest::na.roughfix, and missForest::missForest}
}
\value{
A list with data frames corresponding to the methods used for imputation.
}
\description{
Takes in a geno object and imputes markers using various methods.
}
\section{Methods}{

Markers can be imputed using mean, median, EM, Random Forest (RF), K Nearest Neighbors (KNN), and Singular Value Decomposition (SVD)
}
\examples{
data(wheat)
impute = geno.impute(hap$geno,c("mean","EM"))
impute$mean[1:10,1:20]
impute$EM[1:10,1:20]

}
\author{
Chris Gaynor, \email{chris.gaynor@roslin.ed.ac.uk}

Trevor Rife, \email{trife@ksu.edu}

Jessica Rutkoski, \email{jer263@cornell.edu}
}

