% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.impute.R
\name{hap.impute}
\alias{hap.impute}
\title{Impute markers in gbs}
\usage{
hap.impute(hap, delim, exp.format = c("MEGA4", "FST", "STRUCTURE", "RQTL",
  "AB", "GAPIT"), parents = null)
}
\arguments{
\item{delim}{the deliminator used in the file}

\item{exp.format}{the format you wish to convert the hap object to}

\item{parents}{columns for the two parents being used to convert to RQTL and AB format}

\item{hap.obj}{the hap object to convert}
}
\description{
Takes in hap object and impute markers
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

