\name{kappaii}

\alias{kappaii}

\title{Transformation of the GBSD distribution}

\description{
  This function gives the value of a tranformation of the GBS distribution.
}

\usage{
kappaii(x, theta = c(1.0, 1.0))
}

\arguments{
  \item{x}{Vector of observations.}
  \item{theta}{Vector of parameters of alpha and beta.}
}

\value{
  \code{kappaii()} computes the value of a tranformation of the GBSD.
}

\references{Sanhueza, A., Leiva, V., Balakrishnan, N. (2008) The generalized Birnbaum-Saunders
distribution and its theory, methodology and application. Comm. Stat. Theory and
Meth. 37:645-670.}

\author{
Barros, Michelli <michelli.karinne@gmail.com> \cr 
Leiva, Victor <victor.leiva@uv.cl, victor.leiva@yahoo.com> \cr
Paula, Gilberto A. <giapaula@ime.usp.br> 
}

\keyword{univar}
\keyword{misc}
