% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_process.R
\name{post_process}
\alias{post_process}
\title{Generative Bootstrap Sampler for Logisti Regression Model}
\usage{
post_process(theta, alpha=0.95, level = 3, B1=1000, B2=500, B3=50, B10=50)
}
\arguments{
\item{theta}{The conventional point estimator of interest}
\item{alpha}{The confidence level}
\item{level}{The level of bootstrap}
\item{B1}{The number of bootstrap samples in the first level}

\item{B2}{The number of bootstrap samples in the second level}

\item{B3}{The number of bootstrap samples in the third level}

\item{B10}{The number of the first level bootstrap samples to be evaluated at one time. This is to reduce the amount of RAM and GPU memory in storing the bootstrap samples.}
}
\value{
\item{CI1}{1-alpha confidence interval without iterated bootstrap}
\item{CI2}{1-alpha confidence interval with calibration via double bootstrap}
\item{CI3}{1-alpha studentized confidence interval with calibration via triple bootstrap}
\item{CI_perc}{1-alpha confidence interval via percentile bootstrapping}
\item{T_s}{The bootstrapped ststistics used in constructing "CI2"}
\item{T_s0}{The bootstrapped ststistics used in constructing "CI3"}
\item{d}{The bootstrapped confidence probability for "CI2"}
\item{d0}{The bootstrapped confidence probability for "CI3"}
\item{theta_bc1}{The bias-corrected estimator via the single bootstrap}
\item{theta_bc2}{The bias-corrected estimator via the double bootstrap}
\item{theta_bc3}{The bias-corrected estimator via the triple bootstrap}
\item{alpha}{The confidence level}
\item{time_post}{Computation time for the post processing}
}
\description{
Post-processing procedure to evaluate CIs and bias-corrected estimators.
}
\details{

}
\examples{
#post_process(theta)
}
\seealso{
GBS, GBS_custom, GBS_Sampling
}
\author{
Minsuk Shin and Jun Liu
}
