\name{searchMode}

\alias{searchMode}

\title{Function for obtaining the empirical mode of the data}

\description{
  This function computes the empirical mode of the data.
}

\usage{
searchMode(x)
}

\arguments{
  \item{x}{Vector of observations.}
}

\value{
  code{searchMode()} return the empirical mode (modal value) of the data.
}

\references{}

\author{
Barros, Michelli <michelli.karinne@gmail.com> \cr 
Leiva, Victor <victor.leiva@uv.cl, victor.leiva@yahoo.com> \cr
Paula, Gilberto A. <giapaula@ime.usp.br> 
}

\keyword{univar}
\keyword{math}
