% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.match.R
\name{allele.match}
\alias{allele.match}
\title{Allele match}
\usage{
allele.match(hap, result = c("count", "percent"), genotypes = c(NA, "A",
  "C", "G", "T", "H", "N"), histgraph = F)
}
\arguments{
\item{hap}{a matrix or data frmae consisting consisting of rows (markers) and columns (individuals)}

\item{result}{return the number of matched calls, percent identitiy, or both}

\item{genotypes}{optional vector to include non-standard genotypes}

\item{histgraph}{output histograms of the results}
}
\description{
This function compares alleles across lines in a hap object
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

