% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.match.R
\name{allele.match}
\alias{allele.match}
\title{Allele match}
\usage{
allele.match(hap, result = c("count", "percent"), calls = NULL)
}
\arguments{
\item{hap}{a matrix or data frmae consisting consisting of rows (markers) and columns (individuals)}

\item{result}{return the number of matched calls, percent identitiy, or both}

\item{calls}{optional vector to include non-standard genotype calls}
}
\description{
This function compares alleles across lines in a hap object
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{alleles,}
\keyword{hap}

