% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele.match.R
\name{allele.match}
\alias{allele.match}
\title{Allele match.}
\usage{
allele.match(hap, result = c("count", "percent"), histgraph = F)
}
\arguments{
\item{hap}{A matrix or data frame consisting consisting of rows (markers) and columns (individuals).}

\item{result}{The values to be returned. Either the number of matched calls (\code{count}), \code{percent} identitiy, or both.}

\item{histgraph}{Logical value to graph results.}
}
\value{
A matrix with upper triangle equal to the percent identity of the individuals and lower triangle equal to the proportion of alleles that match between two individuals.
}
\description{
Compares alleles across different individuals in a hap object.
}
\details{
This function ignores missing data and heterozygous calls.
}
\examples{
data(wheat)
wheat.match = allele.match(wheat[,6:ncol(wheat)],histgraph=T)

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}

