% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.summary.R
\name{filter.summary}
\alias{filter.summary}
\title{Filter Summary}
\usage{
filter.summary(hap, project = "gbs", output = "hap", graph = F)
}
\arguments{
\item{hap}{the hap object to be processed}

\item{project}{a project name that will be used to name each tag}

\item{output}{type of output desired. hap will use the same calls, geno will convert the hap object to numeric encoding}

\item{graph}{output graphs for certain stats (blank wells, maf, percent present, het, percent het)}
}
\description{
Processes hap object, calculates summary statistics, returns list with hap and/or geno objects, and prints several graphs of summary statistics
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}

Jesse Poland, \email{jpoland@ksu.edu}
}
\keyword{}

