% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.summary.R
\name{filter.summary}
\alias{filter.summary}
\title{Filter Summary}
\usage{
filter.summary(hap)
}
\arguments{
\item{hap}{the hap object to be processed}
}
\description{
Processes hap object, calculates summary statistics, and graphs relevant stats
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

