% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.dendro.R
\name{gbs.dendro}
\alias{gbs.dendro}
\title{GBS Dendrograms}
\usage{
gbs.dendro(geno, df, taxa, tips, leafs, tipColors, leafColors, ...)
}
\arguments{
\item{geno}{the geno object to manipulate}

\item{df}{data frame containing items to be plotted}

\item{taxa}{string representing the column in the dataframe that contains the names of the taxa in the geno object}

\item{tips}{string representing the column in the dataframe that contains the trait of interest for the tips}

\item{leafs}{string representing the column in the dataframe that contains the trait of interest for the leafs}

\item{...}{additional graphical arguments to pass to the plot.phylo function}

\item{tipColor}{a vector of colors to be used for the tips}

\item{leafColor}{a vector of colors to be used for the leafs}
}
\description{
Create different dendrograms with GBS data borrows some code from picante package
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}

Narinder Singh, \email{nss470@ksu.edu}
}
\keyword{}

