% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.dendro.R
\name{gbs.dendro}
\alias{gbs.dendro}
\title{GBS Dendrograms}
\usage{
gbs.dendro(hap)
}
\arguments{
\item{hap}{the hap object to manipulate}
}
\description{
Create different dendrograms with GBS data
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}

Narinder Singh, \email{nss470@ksu.edu}
}
\keyword{}

