% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.diversity.R
\name{gbs.diversity}
\alias{gbs.diversity}
\title{GBS diversity.}
\usage{
gbs.diversity(hap, clusters = NULL, maf.thresh = 0.05, graph = F,
  het = "H", missing = "N")
}
\arguments{
\item{hap}{The gbs object to manipulate.}

\item{clusters}{A data frame with row names equal to individuals and a single column dictating which group each individual belongs to.}

\item{graph}{A logical option to graph the percent polymorphism.}

\item{het}{The symbol(s) used for heterozygous calls.}

\item{missing}{The symbol(s) used for missing calls.}
}
\value{
A list containing the Nei's genetic diversity and FST of the individuals
}
\description{
Calculates different measures of diveristy in a gbs object.
}
\details{

}
\examples{
data(wheat)
gbs.diversity(hap,graph=T)

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}

