% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.diversity.R
\name{gbs.diversity}
\alias{gbs.diversity}
\title{Determine GBS diversity.}
\usage{
gbs.diversity(hap, clusters = NULL, maf.thresh = 0.05, graph = F)
}
\arguments{
\item{hap}{The hap object to manipulate.}

\item{clusters}{A data frame with row names equal to individuals and a single column dictating which group each individual belongs to.}

\item{graph}{Logical option to graph percent polymorphism.}
}
\value{
A list containing the Nei's genetic diversity and FST of the individuals
}
\description{
Calculates different measures of diveristy in a hap object.
}
\details{

}
\examples{
data(wheat)
hap = hap.read(wheat)
hap = filter.summary(hap,project="wheat",output="hap")$hap
gbs.diversity(hap,data.col=14,graph=T)

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}

