% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.diversity.R
\name{gbs.diversity}
\alias{gbs.diversity}
\title{GBS Diversity}
\usage{
gbs.diversity(geno)
}
\arguments{
\item{hap}{the hap object to manipulate}

\item{delim}{the delimiter for the hap object (if file)}
}
\description{
Calculate different measures of diveristy in a hap object
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

