% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.diversity.R
\name{gbs.diversity}
\alias{gbs.diversity}
\title{GBS Diversity}
\usage{
gbs.diversity(hap, data.col, maf.thresh = 0, graph = F)
}
\arguments{
\item{hap}{the hap object to manipulate}

\item{data.col}{the column number with the first individual}

\item{maf.thresh}{threshold for polymorphism}

\item{graph}{option to graph percent polymorphism}
}
\description{
Calculate different measures of diveristy in a hap object
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

