% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.diversity.R
\name{gbs.diversity}
\alias{gbs.diversity}
\title{GBS Diversity}
\usage{
gbs.diversity(hap)
}
\arguments{
\item{hap}{the hap object to manipulate}
}
\description{
Calculate different measures of diveristy in a hap object
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

