% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.graph.R
\name{gbs.graph}
\alias{gbs.graph}
\title{GBS Graphs}
\usage{
gbs.graph(hap)
}
\arguments{
\item{hap}{the hap object to be processed}
}
\description{
Calculates and graphs relevant stats for a hap object
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

