% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbs.summary.R
\name{gbs.summary}
\alias{gbs.summary}
\title{GBS Summary}
\usage{
gbs.summary(hap.file, delim = "\\t")
}
\arguments{
\item{delim}{the delimiter for the hap file}

\item{hap.obj}{the hap object to convert}
}
\description{
This function processes and filters hap files
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

