% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.collapse.R
\name{hap.collapse}
\alias{hap.collapse}
\title{Merge duplicate individuals.}
\usage{
hap.collapse(hap, names, match, het = "H")
}
\arguments{
\item{hap}{A gbs object consisting only of individuals (columns) and markers (rows).}

\item{names}{A list of the base names of the individual(s) to merge.}

\item{match}{The percent matching individuals must be before merging.}

\item{het}{The symbol(s) used for heterozygous calls.}
}
\value{
A gbs object with duplicate individuals merged.
}
\description{
Combines multiple individuals that have the same name (e.g. ind.1,ind.2,...,ind.n) in a gbs object.
}
\details{
This function merges the calls from duplicate individuals. If a call for a given marker does not match between individuals, it is called heterozygous. All heterozygous markers are stored as "H".
}
\examples{
data(dh)
dh = hap.collapse(dh,names=c("TIGER","DANBY"),match=0.9)
head(dh)

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}

