% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.convert.R
\name{hap.convert}
\alias{hap.convert}
\title{Convert hap objects to other formats}
\usage{
hap.convert(hap, delim, exp.format = c("MEGA4", "FST", "STRUCTURE", "RQTL",
  "AB", "GAPIT"), parents = null)
}
\arguments{
\item{delim}{the deliminator used in the file}

\item{exp.format}{the format you wish to convert the hap object to}

\item{parents}{columns for the two parents being used to convert to RQTL and AB format}

\item{hap.obj}{the hap object to convert}
}
\description{
Takes in hap files and converts them to the formats needed for other analysis programs
}
\examples{

}
\author{
Narinder Singh, \email{nss470@ksu.edu}

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

