% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.convert.R
\name{hap.convert}
\alias{hap.convert}
\title{Convert hap objects to other formats}
\usage{
hap.convert(hap, format = c("MEGA", "STRUCTURE", "FSTRUCTURE", "RQTL", "AB",
  "GAPIT", "JOINMAP"), parents = NULL, jm.pop = c("BC1", "F2", "RIx", "DH",
  "DH1", "DH2", "HAP", "HAP1", "CP", "BCpxFy", "IMxFy"))
}
\arguments{
\item{format}{the format you wish to convert the hap object to}

\item{parents}{columns for the two parents being used to convert to RQTL and AB format}

\item{hap.obj}{the hap object to convert}
}
\description{
Takes in hap files and converts them to the formats needed for other analysis programs
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}

Narinder Singh, \email{nss470@ksu.edu}

Traci Kantarski, \email{tkantarski@ksu.edu}
}
\keyword{}

