% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.impute.R
\name{hap.impute}
\alias{hap.impute}
\title{Impute markers in gbs}
\usage{
hap.impute(geno, method = c("mean", "EM", "RF", "HMM", "median", "KNN",
  "SVD"))
}
\arguments{
\item{geno}{the geno object to impute}

\item{method}{the method to use for imputation}

\item{parents}{columns for the two parents being used to convert to RQTL and AB format}
}
\description{
Takes in hap object and impute markers
}
\examples{

}
\author{
Chris Gaynor

Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

