% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.impute.R
\name{hap.impute}
\alias{hap.impute}
\title{Impute GBS markers.}
\usage{
hap.impute(geno, method, k = 4, maxiter = 10, n.core = 1, ...)
}
\arguments{
\item{geno}{The geno object to impute. Each row is an individual and each column is a marker. Genotypic data is expected to be encoded as \code{-1,0,1}.}

\item{method}{The method to use for imputation. See details.}

\item{k}{The number of clusters for KNN imputation.}

\item{maxiter}{The maximum number of iterations for RF and SVD imputation.}

\item{n.core}{The number of cores to use for RF imputation.}

\item{...}{Aadditional arguments for rrBLUP::A.mat, randomForest::na.roughfix, missForest::missForest}
}
\value{

}
\description{
Takes in a geno object and imputes markers using various methods.
}
\section{Methods}{

Markers can be imputed using mean, median, EM, Random Forest (RF), K Nearest Neighbors (KNN), and Singular Value Decomposition (SVD)
}
\examples{
data(wheat)
hap = hap.read(wheat)
geno = filter.summary(hap,project="wheat",output=c("geno","hap"))$geno
geno.impute = hap.impute(geno,c("MEAN","EM"))
head(geno.impute$mean)
head(geno.impute$EM)

}
\author{
Chris Gaynor, \email{chris.gaynor@roslin.ed.ac.uk}

Trevor Rife, \email{trife@ksu.edu}

Jessica Rutkoski, \email{jer263@cornell.edu}
}

