% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.join.R
\name{hap.join}
\alias{hap.join}
\title{Hap Join}
\usage{
hap.join(hap.dir, delim = "\\t")
}
\arguments{
\item{hap.dir}{directory where hap files are located}
}
\description{
Combines multiple hap files into a single hap object
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\keyword{}

