% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.join.R
\name{hap.join}
\alias{hap.join}
\title{Bind multiple hap files together}
\usage{
hap.join(hap.dir, delim = "\\t")
}
\arguments{
\item{hap.dir}{The directory where the hap files are located.}

\item{delim}{The file delimiter for the hap files.}
}
\description{
Combines multiple hap files into a single hap object.
}
\details{
For use with the old TASSEL 4 GBS pipeline that produced multiple files
}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\seealso{
\code{\link{hap.read}}
}

