% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.read.R
\name{hap.read}
\alias{hap.read}
\title{Read hap file(s) or data frame.}
\usage{
hap.read(hap.obj, delim = "\\t", data.col)
}
\arguments{
\item{delim}{The delimter used in the file(s).}

\item{data.col}{The column number for the first line or individual.}

\item{hap}{A hap file or folder containing hap file(s).}
}
\value{
A gbs object.
}
\description{
Read a file or data frame to create a gbs object
}
\details{
This function creates a gbs object from hap file(s) or a data frame. The gbs object contains a data frame representing the header columns, a data frame representing the genotypic calls, an empty geno data frame, and an empty stats data frame. Both the geno and stats data frames are created with the \code{gbs.summary} function. If a folder is passed to \code{hap.read}, \code{hap.join} is called to merge all files within the folder and remove duplicate tags.
}
\examples{
data(wheat)
hap = hap.read(raw.hap)

}
\author{
Trevor Rife, \email{trife@ksu.edu}

Narinder Singh, \email{nss470@ksu.edu}
}

