% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.read.R
\name{hap.read}
\alias{hap.read}
\title{Read hap file}
\usage{
hap.read(file, delim = "\\t", data)
}
\arguments{
\item{file}{the hap file}

\item{delim}{the delimter used in the file}

\item{data}{column number where genotypes begin}
}
\description{
Read in a file to create a hap object
}
\examples{

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}
\keyword{hap}

