% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.read.R
\name{hap.read}
\alias{hap.read}
\title{Read hap file(s) or hap object.}
\usage{
hap.read(hap.obj, delim = "\\t", data.col)
}
\arguments{
\item{delim}{The delimter used in the file(s).}

\item{data.col}{The column number for the first line or individual.}

\item{hap}{The hap object, file, or folder containing hap files.}
}
\value{
A gbs object.
}
\description{
Read a file or data frame to create a hap object
}
\details{
If a folder is passed to \code{hap.read}, \code{hap.join} is called to merge all files within the folder and remove duplicates.
}
\examples{
data(wheat)
hap = hap.read(wheat)

}
\author{
Trevor Rife, \email{trife@ksu.edu}
}

